fix: Emoji picker overlay
